with casting

+++ utility.c
@@ -204,7 +204,7 @@ Request *runRequest(ProcessQueue *pq /*Stats *stats*/)
         pthread_cond_wait(&(pq->not_empty), &(pq->mutex));
     }
     Request *request = queuePopHead(pq->waiting_queue);
-    queueInsert(pq->running_queue, request, pthread_self()); // stats->id?
+    queueInsert(pq->running_queue, request, (int)(unsigned long) pthread_self()); // stats->id?
 
     // struct timeval end;
     // gettimeofday(&end, NULL);
@@ -218,7 +218,7 @@ Request *runRequest(ProcessQueue *pq /*Stats *stats*/)
 void removeRequest(ProcessQueue *pq, int thread_id)
 {
     pthread_mutex_lock(&(pq->mutex));
-    Request *request = queueRemoveById(pq->running_queue, pthread_self());
+    Request *request = queueRemoveById(pq->running_queue, (int)(unsigned long) pthread_self());
     if (request)
     {
         close(request->connfd);
